+++ ui/app/pages/routes/index.js
@@ -376,16 +376,8 @@ class Routes extends Component {
 
     if (providerName === 'mainnet') {
       name = this.context.t('connectingToMainnet')
-    } else if (providerName === 'ropsten') {
-      name = this.context.t('connectingToRopsten')
-    } else if (providerName === 'kovan') {
-      name = this.context.t('connectingToKovan')
-    } else if (providerName === 'rinkeby') {
-      name = this.context.t('connectingToRinkeby')
-    } else if (providerName === 'localhost') {
-      name = this.context.t('connectingToLocalhost')
-    } else if (providerName === 'goerli') {
-      name = this.context.t('connectingToGoerli')
+    } else if (providerName === 'testnet') {
+      name = this.context.t('connectingToTestnet')
     } else {
       name = this.context.t('connectingTo', [providerId])
     }
@@ -401,16 +393,10 @@ class Routes extends Component {
 
     if (providerName === 'mainnet') {
       name = this.context.t('mainnet')
-    } else if (providerName === 'ropsten') {
-      name = this.context.t('ropsten')
-    } else if (providerName === 'kovan') {
-      name = this.context.t('kovan')
-    } else if (providerName === 'rinkeby') {
-      name = this.context.t('rinkeby')
+    } else if (providerName === 'testnet') {
+      name = this.context.t('testnet')
     } else if (providerName === 'localhost') {
       name = this.context.t('localhost')
-    } else if (providerName === 'goerli') {
-      name = this.context.t('goerli')
     } else {
       name = this.context.t('unknownNetwork')
     }